# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Marketplace Booking API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read Admin Users",
        "operationId": "read_admin_users_admin_users_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserRead"
                  },
                  "title": "Response Read Admin Users Admin Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/listings/pricing-scope-summary": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read Pricing Scope Counts",
        "operationId": "read_pricing_scope_counts_admin_listings_pricing_scope_summary_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingPricingScopeCount"
                  },
                  "title": "Response Read Pricing Scope Counts Admin Listings Pricing Scope Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/listings/promoted": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read Promoted Listings",
        "operationId": "read_promoted_listings_admin_listings_promoted_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingPromotionDetail"
                  },
                  "title": "Response Read Promoted Listings Admin Listings Promoted Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/listings/promotions/summary": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read Promoted Summary",
        "operationId": "read_promoted_summary_admin_listings_promotions_summary_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingPromotionSummary"
                  },
                  "title": "Response Read Promoted Summary Admin Listings Promotions Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/listings/promotions/events": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read Promotion Events",
        "operationId": "read_promotion_events_admin_listings_promotions_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingPromotionEvent"
                  },
                  "title": "Response Read Promotion Events Admin Listings Promotions Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/listings/{listing_id}/promotion": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Listing Promotion",
        "operationId": "update_listing_promotion_admin_listings__listing_id__promotion_patch",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Listing Id"
            }
          },
          {
            "name": "promoted",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Promoted"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/platform-fees/history": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read Platform Fee History",
        "operationId": "read_platform_fee_history_admin_platform_fees_history_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformFeeHistoryPoint"
                  },
                  "title": "Response Read Platform Fee History Admin Platform Fees History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminUserRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "AdminUserRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListingImageRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "alt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Text"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "listing_id",
          "image_url",
          "created_at"
        ],
        "title": "ListingImageRead"
      },
...
      "ListingPricingScopeCount": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "count"
        ],
        "title": "ListingPricingScopeCount"
      },
      "ListingPromotionDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "seller_id"
        ],
        "title": "ListingPromotionDetail"
      },
      "ListingPromotionEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "promoted": {
            "type": "boolean",
            "title": "Promoted"
          },
          "platform_fee_rate": {
            "type": "string",
            "title": "Platform Fee Rate"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "listing_id",
          "seller_id",
          "promoted",
          "platform_fee_rate",
          "created_at"
        ],
        "title": "ListingPromotionEvent"
      },
      "ListingPromotionSummary": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "type",
          "count"
        ],
        "title": "ListingPromotionSummary"
      },
      "ListingRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "type": {
            "$ref": "#/components/schemas/ListingType"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "price_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "USD"
          },
          "inventory_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inventory Count"
          },
          "requires_booking": {
            "type": "boolean",
            "title": "Requires Booking",
            "default": false
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "is_local_only": {
            "type": "boolean",
            "title": "Is Local Only",
            "default": true
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "pickup_enabled": {
            "type": "boolean",
            "title": "Pickup Enabled",
            "default": false
          },
          "meetup_enabled": {
            "type": "boolean",
            "title": "Meetup Enabled",
            "default": false
          },
          "delivery_enabled": {
            "type": "boolean",
            "title": "Delivery Enabled",
            "default": false
          },
          "shipping_enabled": {
            "type": "boolean",
            "title": "Shipping Enabled",
            "default": false
          },
          "lead_time_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Time Hours"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ListingImageRead"
            },
            "type": "array",
            "title": "Images",
            "default": []
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "last_operating_adjustment_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Operating Adjustment At"
          },
          "last_operating_adjustment_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Operating Adjustment Summary"
          },
          "last_pricing_comparison_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Pricing Comparison Scope"
          },
          "available_today": {
            "type": "boolean",
            "title": "Available Today",
            "default": false
          },
          "is_new_listing": {
            "type": "boolean",
            "title": "Is New Listing",
            "default": false
          },
          "recent_transaction_count": {
            "type": "integer",
            "title": "Recent Transaction Count",
            "default": 0
          },
          "is_promoted": {
            "type": "boolean",
            "title": "Is Promoted",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "seller_id",
          "title",
          "slug",
          "type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ListingRead"
      },
      "ListingType": {
        "type": "string",
        "enum": [
          "product",
          "service",
          "hybrid"
        ],
        "title": "ListingType"
      },
...
      "PlatformFeeHistoryPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "order_fee_cents": {
            "type": "integer",
            "title": "Order Fee Cents"
          },
          "booking_fee_cents": {
            "type": "integer",
            "title": "Booking Fee Cents"
          }
        },
        "type": "object",
        "required": [
          "date",
          "order_fee_cents",
          "booking_fee_cents"
        ],
        "title": "PlatformFeeHistoryPoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}